# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casa de Apuestas API",
        "description": "API para casa de apuestas: autenticación, usuarios, deportes, eventos, cuotas, wallet y apuestas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API versión 1"
        }
    ],
    "paths": {
...
        "/events/{sportEvent}/odds": {
            "get": {
                "tags": [
                    "Odds"
                ],
                "summary": "Listar cuotas disponibles de un evento",
                "description": "Devuelve cuotas activas agrupadas por mercado y bookmaker para usuarios autenticados.",
                "operationId": "9b267f38643693f2022f2f330b8d1e73",
                "parameters": [
                    {
                        "name": "sportEvent",
                        "in": "path",
                        "description": "ID interno del evento deportivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuotas disponibles obtenidas correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/odds/sync": {
            "post": {
                "tags": [
                    "Admin Odds"
                ],
                "summary": "Sincronizar cuotas desde The Odds API",
                "description": "Consulta cuotas por deporte, región y mercado; guarda snapshots y detecta cambios por hash.",
                "operationId": "0e6454998b59259290165bb219baf64e",
                "parameters": [
                    {
                        "name": "sport_key",
                        "in": "query",
                        "description": "Filtrar eventos por deporte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "soccer_australia_aleague"
                    },
                    {
                        "name": "event_id",
                        "in": "query",
                        "description": "ID interno del evento específico",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "regions",
                        "in": "query",
                        "description": "Regiones de bookmakers",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "au"
                    },
                    {
                        "name": "markets",
                        "in": "query",
                        "description": "Mercados de apuesta",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "h2h"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Cantidad máxima de eventos a procesar",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuotas sincronizadas correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al sincronizar cuotas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sports/active": {
            "get": {
                "tags": [
                    "Sports"
                ],
                "summary": "Listar deportes activos",
                "description": "Devuelve los deportes activos disponibles para usuarios autenticados.",
                "operationId": "5369b278c4142ada5d56815afe4d99fe",
                "responses": {
                    "200": {
                        "description": "Deportes activos obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sports": {
            "get": {
                "tags": [
                    "Admin Sports"
                ],
                "summary": "Listar deportes para administración",
                "description": "Devuelve deportes paginados para administración, con filtros opcionales.",
                "operationId": "d90dbc5eeea7b8273c215ae47331229a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por título, sport_key o grupo",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "soccer"
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrar por estado activo/inactivo",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 15
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deportes obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sports/sync": {
            "post": {
                "tags": [
                    "Admin Sports"
                ],
                "summary": "Sincronizar deportes desde The Odds API",
                "description": "Consulta The Odds API y guarda o actualiza deportes sin duplicarlos.",
                "operationId": "1e1b4d998b3635bc00258462a32f54b3",
                "responses": {
                    "200": {
                        "description": "Deportes sincronizados correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error al sincronizar deportes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sports/{sport}/status": {
            "patch": {
                "tags": [
                    "Admin Sports"
                ],
                "summary": "Activar o desactivar deporte",
                "description": "Permite a un administrador activar o desactivar un deporte.",
                "operationId": "53252d5613f019a33414181316dae75b",
                "parameters": [
                    {
                        "name": "sport",
                        "in": "path",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "active"
                                ],
                                "properties": {
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado del deporte actualizado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token personal de acceso. Formato: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}